# docs/swagger.yaml
basePath: /
definitions:
  dto.AdjustStockRequest:
    properties:
      quantity:
        type: number
      reason:
        type: string
    required:
    - quantity
    - reason
    type: object
...
  dto.ApplicationResponse:
    properties:
      application_date:
        type: string
      application_type:
        $ref: '#/definitions/entities.ApplicationType'
      created_at:
        type: string
      created_by:
        type: integer
      dose_per_ha:
        type: number
      equipment:
        type: string
      field_id:
        type: integer
      field_name:
        type: string
      humidity:
        type: number
      id:
        type: integer
      input_id:
        type: integer
      input_name:
        type: string
      notes:
        type: string
      operator:
        type: string
      planting_id:
        type: integer
      spray_volume:
        type: number
      target:
        type: string
      temperature:
        type: number
      total_used:
        type: number
      wind_speed:
        type: number
    type: object
...
  dto.CreateApplicationRequest:
    properties:
      application_date:
        type: string
      application_type:
        allOf:
        - $ref: '#/definitions/entities.ApplicationType'
        enum:
        - fertilizer
        - herbicide
        - fungicide
        - insecticide
        - correctant
        - biological
      dose_per_ha:
        type: number
      equipment:
        type: string
      field_id:
        type: integer
      humidity:
        maximum: 100
        minimum: 0
        type: number
      input_id:
        type: integer
      notes:
        type: string
      operator:
        type: string
      planting_id:
        type: integer
      spray_volume:
        type: number
      target:
        type: string
      temperature:
        type: number
      wind_speed:
        minimum: 0
        type: number
    required:
    - application_date
    - application_type
    - dose_per_ha
    - field_id
    - input_id
    - operator
    type: object
...
  dto.CreateInputRequest:
    properties:
      batch_number:
        type: string
      category:
        allOf:
        - $ref: '#/definitions/entities.InputCategory'
        enum:
        - fertilizer
        - herbicide
        - fungicide
        - insecticide
        - correctant
        - biological
        - seed
        - other
      cost_per_unit:
        minimum: 0
        type: number
      expiration_date:
        type: string
      manufacturer:
        type: string
      min_stock_qty:
        minimum: 0
        type: number
      name:
        maxLength: 150
        minLength: 2
        type: string
      stock_qty:
        minimum: 0
        type: number
      unit:
        type: string
    required:
    - category
    - name
    - unit
    type: object
...
  dto.CreatePlantingRequest:
    properties:
      crop_id:
        type: integer
      density_kg_ha:
        type: number
      depth_cm:
        type: number
      expected_harvest:
        type: string
      field_id:
        type: integer
      notes:
        type: string
      planting_date:
        type: string
      responsible:
        type: string
      season_id:
        type: integer
      seeds_used_kg:
        type: number
      spacing:
        type: number
    required:
    - crop_id
    - field_id
    - planting_date
    - season_id
    type: object
...
  dto.InputResponse:
    properties:
      active:
        type: boolean
      batch_number:
        type: string
      category:
        $ref: '#/definitions/entities.InputCategory'
      cost_per_unit:
        type: number
      created_at:
        type: string
      expiration_date:
        type: string
      id:
        type: integer
      low_stock:
        type: boolean
      manufacturer:
        type: string
      min_stock_qty:
        type: number
      name:
        type: string
      stock_qty:
        type: number
      unit:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.PaginatedResponse:
    properties:
      data: {}
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  dto.PlantingResponse:
    properties:
      created_at:
        type: string
      created_by:
        type: integer
      crop_id:
        type: integer
      crop_name:
        type: string
      density_kg_ha:
        type: number
      depth_cm:
        type: number
      expected_harvest:
        type: string
      field_id:
        type: integer
      field_name:
        type: string
      id:
        type: integer
      notes:
        type: string
      planting_date:
        type: string
      responsible:
        type: string
      season_id:
        type: integer
      season_name:
        type: string
      seeds_used_kg:
        type: number
      spacing:
        type: number
      status:
        $ref: '#/definitions/entities.PlantingStatus'
      updated_at:
        type: string
    type: object
...
  dto.UpdateInputRequest:
    properties:
      active:
        type: boolean
      batch_number:
        type: string
      cost_per_unit:
        minimum: 0
        type: number
      expiration_date:
        type: string
      manufacturer:
        type: string
      min_stock_qty:
        minimum: 0
        type: number
      name:
        maxLength: 150
        minLength: 2
        type: string
      unit:
        type: string
    required:
    - name
    - unit
    type: object
  dto.UpdatePlantingRequest:
    properties:
      expected_harvest:
        type: string
      notes:
        type: string
      responsible:
        type: string
      status:
        allOf:
        - $ref: '#/definitions/entities.PlantingStatus'
        enum:
        - active
        - harvested
        - lost
    type: object
...
  entities.ApplicationType:
    enum:
    - fertilizer
    - herbicide
    - fungicide
    - insecticide
    - correctant
    - biological
    type: string
    x-enum-varnames:
    - AppTypeFertilizer
    - AppTypeHerbicide
    - AppTypeFungicide
    - AppTypeInsecticide
    - AppTypeCorrectant
    - AppTypeBiological
...
  entities.InputCategory:
    enum:
    - fertilizer
    - herbicide
    - fungicide
    - insecticide
    - correctant
    - biological
    - seed
    - other
    type: string
    x-enum-varnames:
    - InputCategoryFertilizer
    - InputCategoryHerbicide
    - InputCategoryFungicide
    - InputCategoryInsecticide
    - InputCategoryCorrectant
    - InputCategoryBiological
    - InputCategorySeed
    - InputCategoryOther
...
  entities.PlantingStatus:
    enum:
    - active
    - harvested
    - lost
    type: string
    x-enum-varnames:
    - PlantingStatusActive
    - PlantingStatusHarvested
    - PlantingStatusLost
...
  utils.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API REST de gestão agrícola com autenticação JWT
  title: AgroControl API
  version: "1.0"
paths:
...
  /api/v1/applications:
    get:
      parameters:
      - description: Página
        in: query
        name: page
        type: integer
      - description: Itens por página
        in: query
        name: limit
        type: integer
      - description: Filtrar por talhão
        in: query
        name: field_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedResponse'
      security:
      - BearerAuth: []
      summary: Listar aplicações
      tags:
      - applications
    post:
      consumes:
      - application/json
      description: Registra uma aplicação de insumo em um talhão. Debita o estoque
        automaticamente e cria alerta se estoque ficar baixo.
      parameters:
      - description: Dados da aplicação
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateApplicationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ApplicationResponse'
        "422":
          description: Estoque insuficiente ou talhão inativo
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Registrar aplicação de insumo
      tags:
      - applications
  /api/v1/applications/{id}:
    get:
      parameters:
      - description: ID da aplicação
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApplicationResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Buscar aplicação por ID
      tags:
      - applications
...
  /api/v1/fields/{id}/applications:
    get:
      parameters:
      - description: ID do talhão
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ApplicationResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Listar aplicações por talhão
      tags:
      - applications
...
  /api/v1/inputs:
    get:
      description: Retorna insumos paginados com filtro por categoria
      parameters:
      - description: Página
        in: query
        name: page
        type: integer
      - description: Itens por página
        in: query
        name: limit
        type: integer
      - description: Busca por nome
        in: query
        name: search
        type: string
      - description: Filtro por categoria (fertilizer, herbicide, fungicide...)
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedResponse'
      security:
      - BearerAuth: []
      summary: Listar insumos
      tags:
      - inputs
    post:
      consumes:
      - application/json
      description: Cadastra um novo insumo agrícola com controle de estoque
      parameters:
      - description: Dados do insumo
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateInputRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.InputResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Criar insumo
      tags:
      - inputs
  /api/v1/inputs/{id}:
    delete:
      parameters:
      - description: ID do insumo
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Excluir insumo
      tags:
      - inputs
    get:
      parameters:
      - description: ID do insumo
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.InputResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Buscar insumo por ID
      tags:
      - inputs
    put:
      consumes:
      - application/json
      parameters:
      - description: ID do insumo
        in: path
        name: id
        required: true
        type: integer
      - description: Dados atualizados
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateInputRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.InputResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Atualizar insumo
      tags:
      - inputs
  /api/v1/inputs/{id}/stock:
    patch:
      consumes:
      - application/json
      description: Ajusta o estoque do insumo. Positivo = entrada, negativo = saída
        manual.
      parameters:
      - description: ID do insumo
        in: path
        name: id
        required: true
        type: integer
      - description: Quantidade a ajustar
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AdjustStockRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.InputResponse'
        "422":
          description: Estoque insuficiente
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Ajustar estoque
      tags:
      - inputs
...
  /api/v1/plantings:
    get:
      description: Admin vê todos os plantios, outros veem apenas os seus
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PlantingResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar plantios
      tags:
      - plantings
    post:
      consumes:
      - application/json
      description: Registra um novo plantio em um talhão ativo. Valida se o talhão
        está ativo e se o usuário tem acesso.
      parameters:
      - description: Dados do plantio
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePlantingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.PlantingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "422":
          description: Talhão inativo
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Criar plantio
      tags:
      - plantings
  /api/v1/plantings/{id}:
    delete:
      parameters:
      - description: ID do plantio
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Excluir plantio
      tags:
      - plantings
    get:
      parameters:
      - description: ID do plantio
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PlantingResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Buscar plantio por ID
      tags:
      - plantings
    put:
      consumes:
      - application/json
      parameters:
      - description: ID do plantio
        in: path
        name: id
        required: true
        type: integer
      - description: Dados atualizados
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePlantingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PlantingResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Atualizar plantio
      tags:
      - plantings
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"